Temporary commit for testing
In 9005450, it appears as though those command-line options for which "allow_multiple" is set to true default to being visible, despite "visible" being set to false. In this commit, the "allow_multiple" flag is removed; if this now prevents such command-line options to defaulting to being visible, that would suggest that there may be some parsing error within Giraffe.